fix(api): startup fixes for Brand, Collection, Category, Product and Stock modules. Auto update for user-role-rule relations

+++ src/main/kotlin/siberia/modules/product/data/dto/ProductCreateDto.kt
@@ -18,7 +18,7 @@ data class ProductCreateDto (
     val collection: Int? = null,
     val color: String,
     val amountInBox: Int,
-    val expirationDate: Int,
+    val expirationDate: Long,
     val link: String
 
 //    Future iterations
