+++ src/main/kotlin/siberia/modules/product/data/dto/ProductOutputDto.kt
@@ -21,7 +21,7 @@ data class ProductOutputDto(
     val collection: Int?,
     val color: String,
     val amountInBox: Int,
-    val expirationDate: Int,
+    val expirationDate: Long,
     val link: String,
 
 //    Future iterations
